更新

+++ server_code/gameworld/script/interface/redis_api.h
@@ -1,13 +1,13 @@
 #ifndef _REDIS_API_H_
 #define _REDIS_API_H_
+#include "StdAfx.h"
 /*
 
 */
-
 //tolua_begin
 namespace LRedis
 {
-	void Zadd(char* rankname,int score,char* actordata);
+	void Zadd(char* rankname, long long score, char* actordata);
 }
 //tolua_end
